Add environment support to discovery
Previous to this commit, the only way to discover nodes within a given
environment is to use a fact value (and have a custom environment fact).
This commit adds the *environment* key to the discover custom function
allowing users to specify an environment to limit the discovery to.  It
still requires a custom environment fact