Load classes, not files
Previous to this commit, the orchestrate Puppet application loaded files
with the .mc extension passed to the command as a command line argument.
This command removes that functionality in favor of receiving a Puppet
class instead of a file.  The orchestrate command runs the appropriate
Puppet class from the specified environment (defaults to production),
compiles a catalog, and runs the orchestration script.